Removed header bar, buttons/elements moved into main layout
