+++ interface/src/application/pages/DevicePages/SettingsPage/InfoBlock.tsx
