+++ interface/src/application/pages/DevicePages/SettingsPage/KinematicsBlock.tsx
