+++ interface/src/application/pages/DevicePages/SettingsPage/KinematicsDiagram/delta-dark.png
