+++ interface/src/application/pages/DevicePages/SettingsPage/KinematicsDiagram/delta-light.png
