+++ interface/src/application/pages/DevicePages/SettingsPage/KinematicsDiagram/index.tsx
