+++ interface/src/application/pages/DevicePages/Views/SummaryTags/CPUTag.tsx
