+++ interface/src/application/pages/DevicePages/Views/SummaryTags/TemperatureTag.tsx
